Added stub for unmanaged object dynamic api
